fair_queue: enhance the fair queue with rate limits
Currently the fair queue operates by setting a fixed capacity. We will
make sure that the fair queue never allows more requests to be inflight
at the same time than that.

This approach has worked well but has a variety of shortcomings. Chief
among them, it is hard to account for requests of different sizes since
the queue capacity is usually obtained for a particular request size,
and real life workloads will generate a broad distribution of sizes.

We know from Little's law that for any ergodic closed system:

  N = throughput * latency

Disk devices will usually have a maximum IOPS and a maximum bytes/s
that they can handle. Since we expect Little's law to be valid for both
quantities, all we need to do is to find out what is the maximum
expected value for those quantities, and control for that.

The challenge in this patchset is to make this change in a backwards
compatible way, for the benefit of systems that may already be tuning
the queue capacity. To do that, we will allow any of the following three
parameters to be specified:

 - inflight requests (capacity)
 - requests per second
 - bytes per second

The fair queue will allow requests to pass through if none of those
limits are violated, and block if any of them are. If not specified
those quantities will be set to their maximum and they will be
effectively disabled.

It is expected that new systems will control for bytes/s and requests
per second only.

Signed-off-by: Glauber Costa <glauber@scylladb.com>