reactor: allow to control I/O based on IOPS and bandwidth
This patch introduces two new parameters:
 --max-disk-bytes-per-sec
 --max-disk-req-per-sec

The expected format of the parameter is [device:]value in the near
future, but in this patch we will not accept specification of a device
yet.

When we do, we should also be able to accept multiple invocations of
this parameter and then proceed to create one I/O queue per specified
device.

I have asked myself the question of whether we should still allow those
values to be specified externally or if we should just evaluate them
during seastar start.

I have decided for the former, and this is my list of reasons:
 - No matter how quickly we can do the evaluation, that will never be as
   quick as not evaluating. A lot of disks have those parameters very
   well behaved and some users will be justified in preferring the
   evaluation not to run at every start.

 - evaluation accuracy is usually improved by large file sizes due to
   in-disk caches. Generating big files mean the evaluation takes a long
   time. Generating small files mean the evaluation loses reliability.

 - the proposed format still allows evaluation-at-startup to be plugged
   seamlessly if we so wish, by specifying a device name but no
   pre-defined quantity.

For a complete offering we should upgrade iotune to be able to generate
those parameters. However, this patch is already useful as is, since
unlike the iodepth, which is an indirect variable we need to measure,
IOPS and bandwidth are published specs of most disks and users can
already benefit from this infrastructure right now if they trust their
disk specs enough.

Signed-off-by: Glauber Costa <glauber@scylladb.com>