resource: improve io_queue distributing code to account for grouping
Right now create as many I/O queues as the user requested and associate
shards to I/O queues according to what hwloc determines to be best.

In preparation for the upcoming changes in the I/O Scheduler queue
distribution, improve this code so that it can also generate results
that are usable by the grouping code.

The distribution part stays exactly the same: we still want hwloc to
distribute queues fairly around the system and add shards to groups
according to their proximity. However, as we won't have a specified
number of I/O queues anymore we need to calculate a good amount of
groups ourselves.

The code as is after this patch works well for both methods of I/O
Queue creation and distribution.

Signed-off-by: Glauber Costa <glauber@scylladb.com>