io_queue: store io_queue pointer at allocation time
Right now we are allocating a naked pointer in alloc_io_queue and later
moving it to its unique_ptr place holder.

This was originally done because we need the pointer to be available in
a global vector for the assignment phase. However, we can easily do
the allocation using make_unique during alloc_io_queue -> because at
that point we are already in the correct shard - and just assign to
the io_queue pointer vector using unique_ptr::get.

This is cleaner and easier to read.

Signed-off-by: Glauber Costa <glauber@scylladb.com>