io_queue: accept a configuration struct
Will do that so we can add more optional parameters soon. The config
struct is converted to the fair_queue's config struct and we'll start
using that interface.

While doing that: the capacity of the queue is now unsigned everywhere,
but reactor::max_aio is still size_t. Move that to unsigned as well so
we don't have to keep playing the casting game.

Signed-off-by: Glauber Costa <glauber@scylladb.com>