reactor: allow submit_io to take any pointer
We are currently passing a pointer to promise<io_event> to submit_io.  I
would like to enhance the information that is passed down to include a
bit more information about the I/O: whether or not it is a read or a
write, its size, etc.

We can add a structure containing all that to the signature of
submit_io, but I'd like to keep that as an implementation detail inside
reactor.cc.

For that reason, I am changing the signature of submit_io to take a
template Pointer.

Signed-off-by: Glauber Costa <glauber@scylladb.com>