reactor: keep track of size and type of requests
Requests are not all equal and can be of different sizes and types.
Sizes are just number of bytes, and types will be defined by the fair
queue as "input" and "output" (reads and writes).

Those types are chosen so that the fair queue can be kept as
generic as possible, but I still didn't see a reason to allow for type
registration: just two seems to me like it will work just fine.

Once we track those requests at creation time and add them to the I/O
descriptor, we can then notify the fair queue of their completion with
full information. This will be used to implement per-type limits as well
of separate throughput/bandwidth limits.

Signed-off-by: Glauber Costa <glauber@scylladb.com>