io_tester: only create scheduling groups once per all shards
scheduling groups exist once for all shards and are instantiated
individually. The current code was creating a new one in every shard,
which is pointless, because they don't really cross a shard barrier
unlike I/O classes (for I/O classes, we can have multiple shards
dispatching from the same queue so they can't be the same).

The result of this is that running io_tester in a large number
of shards will break, since we don't have enough slots for that
many groups.

Signed-off-by: Glauber Costa <glauber@scylladb.com>