utils: allow parse_memory_size to interpret unitless numbers
I am planning to reuse this interface to parse numbers coming from
disk IOPS and throuhgput numbers. Some disks can be quite slow, and
have less than 1k IOPS (although that is quite rare).

We can trivially adjust the interface so that if a number without
units is passed, it is interpreted as is.

Signed-off-by: Glauber Costa <glauber@scylladb.com>