thread_cputime_clock: add a clock interface for CLOCK_THREAD_CPUTIME_ID
clock_gettime, if called with CLOCK_THREAD_CPUTIME_ID will give us an
idea of how much time the current thread consumed. Using this can be
useful for us to gain visibility of what's happening in the system.

To make the integration with the rest of the code better, we can wrap
it in an interface similar to std::chrono's clocks.

Signed-off-by: Glauber Costa <glauber@scylladb.com>