iotune: generate config file for more than one directory
Our I/O Scheduler can now support more than one mountpoint. So we will
change iotune accordingly. It can now take more than one evaluation
directory, and it will generate data for all of the tested directories
in the yaml file.

The number of I/O queues is chosen so as to be able to uphold the
minimums for the slowest device. This would work to unleash some use
cases. Long term, we are already planning to get rid of the I/O queues
and dispatch from all shards.

Signed-off-by: Glauber Costa <glauber@scylladb.com>