properly configure I/O Scheduler when --max-io-requests is passed
We now have scaling between reads and writes as part of the new I/O
Scheduler. However, when a capacity is explicitly set, which means the
old --max-io-requests is used, we end up miscalculating that, and as a
result the I/O Scheduler fails to uphold the shares.

If we are using the legacy config, which didn't differentiate reads from
writes, then we should hard code the ratio to 1.

Fixes #482

Signed-off-by: Glauber Costa <glauber@scylladb.com>