io_queue: pass io_queue pointer to priority class data
We will then be able to get the mountpoint from it. We will also be able
to derive the coordinator() from it, instead of assuming that the shard
executing the code is the coordinator. This latter benefit will come in
handy later on, as we will start creating this structure directly from the
owner shard.

Signed-off-by: Glauber Costa <glauber@scylladb.com>