fair_queue: use a naked pointer instead of a shared pointer for the interface
Right now most of the accesses to the fair queue are mediated through a
coordinator shard and there are never external direct accesses. We want
to change that, and dealing with shared pointers make that hard, since
foreign users will bump the reference.

To make the interfaces safe against foreign access I am changing the
lw_shared_ptr to a naked pointer. The set that used to hold all
shared_pointers is now turned into an array which is pre-initialized
in the creator shard.

Signed-off-by: Glauber Costa <glauber@scylladb.com>