make the fair queue shared, and the io queue local
Right now, all I/O Queue shard goes through a coordinator shard.  We
want to change that, and as a step towards that direction we will make
some structures that currently belong to the I/O Scheduler local to the
shard and only share the core of the I/O Scheduler: the fair_queue.

Another advantage of doing that is that the _priority_classes map is now
shard-local. Unlike the CPU scheduler, the I/O Scheduler may have to
deal in the same scheduler with classes with the same name going through
the same scheduler (for instance compaction-1, compaction-2, for shards
1 and 2 respectively). This prevents us from using the same identifier
for the class in all shards.

But if the list of classes is local to the shard, then we can convert
the identifiers to always be the same for a specific class and therefore
make them safe to pass around. We will still create pointers for all
shards in the fair_queue, but that level is not visible to the user at
all.

This will be done in a follow up patch.

Signed-off-by: Glauber Costa <glauber@scylladb.com>