rework I/O queue registration API and make class registration synchronous
Up until now the I/O Scheduler has been registering its classes lazily,
and the same class in different shards could have different identifiers.
This stems mostly from the fact that unlike the CPU Scheduler, we can
have many instances of the same class from different shards sharing the
scheduler.

However, in the last patch we have made changes that allowed us to
change that. Since the priority_classes set (now an array) is
shard-local, we can register globally per class and make sure that a
class with the same name gets the same ID regardless of the shard that
it registers from.

To guarantee the above we change the registration API so that it returns
a future. Callers will have to be patched so that registration no longer
happens during normal cpp initializers but are instead futurized. The
io_tester, our in-tree user for the I/O classes is patched.

Once the infrastructure above is in place, we can get rid of
find_or_create_class and stop the lazy initialization: classes are now
initialized upon registration, and since they are currently registered
in every shard they are used we are good.

Signed-off-by: Glauber Costa <glauber@scylladb.com>