io_queue: only increment bytes and ops when the request is executed.
These metrics are wrong as they are today and I have been meaning to
fix them for a while. As they are now, they count the amount of ops / bytes
that are queued into the I/O queue. What is really important for the user
is the amount of ops / bytes that are effectively executed.

Signed-off-by: Glauber Costa <glauber@scylladb.com>