SystemTune: tune the clocksource
We would like to be able to enforce that the preferred clocksource
is used (tsc for x86 architectures). Not doing so can cause applications
that rely heavily on timestamps to take a performance penalty.

If the clocksource is available but not used, maybe there is a reason to
do so (for instance, it may not be trusted), in which case the user must
explicitly select that the clock must be set throught the --tune-clock
option.

To do the tuning, we will leverage the newly created System tuner.

Signed-off-by: Glauber Costa <glauber@scylladb.com>