rework the fair_queue tester
In this commit, I am atoning for my past sins by reworking the
fair_queue test.

The fair queue test has a number of problems that we realized were there
but never fixed. Chief among them is the dependency that the tests have
on timing.

And you can see why: we were young and naive, and the fair queue is a
building block that uses time heavily in its exponential decay
algorithm. The tests can, however, be rewritten in a way that are
a lot less dependent on time.

In most tests the decay doesn't matter, and we can use timing constructs
just in the ones that do. For the others we can just defer with later()
instead of sleeping to make sure we'll have the chance to exit to the
reactor. This makes the tests more stable and reliable.

Despite the time-dependency, these tests have been running just fine
and not being a problem in practice for a long time. However I am about
to touch them to add new functionality and started seeing scenarios that
fail due to timing in the new tests. Instead of going through the pain
of fine tuning all over again, best to rework them.

I am also taking the opportunity to modernize these tests according to
what we've been doing lately. So from a broad perspective the changes
are:

 * using seastar::async for readability.
 * not using futures when not needed, if we are not time-dependent
   anymore we can control the execution steps exactly and just rely
   on requests being moved to and from the inflight array.
 * Enhancing test_env to avoid exposing every single member and changing
   the convention to use _member_names.

Another advantage of the rework is, as we're removing the sleeps opting
for a lockstep model instead, the test becomes much faster.

Before:

*** No errors detected

real	0m55.604s
user	0m4.223s
sys	0m6.640s

After:

*** No errors detected

real	0m3.078s
user	0m0.369s
sys	0m0.369s

Signed-off-by: Glauber Costa <glauber@scylladb.com>