fair_queue: allow the fair_queue to timeout requests
Requests that have timed out already are useless from the application
point of view, but still use resources - often times precious resources
- until they are able to be retired.

If a request is sitting in the I/O Scheduler, for example, and it times
out while doing so, it is pointless to send it to the storage device:
we'll waste bandwidth, and time, just to discard the result when it
reaches upper layers.

This patch adds to the fair_queue the ability to time out requests that
are sitting in its internal queues.

The design decision that I took is mainly centered about changing the
signature of the function that is passed to the queue() method. If a
timeout exception happened, it will then be passed as a parameter to
that method. Here's why:

Requests that timeout do not ever get executed, therefore the fair_queue
tracking of bytes and capacity used don't get bumped. While we could
bump it to make handling of both the exceptional and normal path
simpler, this does come at a high cost: useful bandwidth would be denied
to runnable requests until the consumer of the fair_queue acknowledges
this fact.

For that reason, timed out requests are to be handled differently. The
consumer has to propagate the exception upwards, but will not worry
about calling notify_requests_finish. The I/O scheduler is adjusted so
that we now can handle early exceptions-- exceptions that happened
before the request was even dispatched (so far just timeouts), and
exceptions that happened at the consumer level when trying to dispatch
the request (same as before).

Tests are added to make sure the timeouts are triggering correctly.

Signed-off-by: Glauber Costa <glauber@scylladb.com>