core: remove iocb references from I/O dispatchers
Our async I/O mechanism relies on Linux async I/O interface for now,
which in turn relies on sending iocbs down to the Kernel. Because iocbs
are so ingrained in the interface, they go all the way up and the
write_dma and read_dma functions are already aware of it.

However, there are now other interfaces, like io_uring, that rely on
different data structures. To account for that, we will encapsulate the
I/O information like file descriptor, addresses and sizes into a seastar
structure and only transform this into iocbs down at the lowest level,
where we already know which interface to use.

In a follow up patch, we will have to do the same thing with the
io_event structure.

Signed-off-by: Glauber Costa <glauber@scylladb.com>