core: remove io_event mentions from upper layer interfaces
We currently work with exposing an io_event, the return value of
the Linux async API at layers as high as read/write_dma. There are
now other mechanisms to initiate Async I/O in Linux like io_uring
that works with different structures.

This patch changes the interfaces so we work with size_t types for
as long as possible, and leave io_event as an implementation detail
for the I/O subsystem.

Signed-off-by: Glauber Costa <glauber@scylladb.com>