core: unify I/O pollers
We currently have two different pollers for I/O: one for
submitting I/O, and another for fetching the results of
a previous submission.

There is no particular reason to have this separate at this
stage, and the reasons for a two-poller system seems mostly
historical to me (at some point, we could submit I/O outside
of polling context)

One of the pollers (getevents) is available - as far as ifdefs
are concerned - on OSv, but I believe that even if this is not
an overlook, if we haven't dispatched anything we'll hardly have
anything to read so I am just ignoring that.

Signed-off-by: Glauber Costa <glauber@scylladb.com>