io_queue: return number of requests dispatched.
Right now we ignore the number of requests dispatched by the I/O queue.

Dispatching happens inside a seastar poller, but that poller does other
things as well and that is enough to determine whether or not the poller
did work.

I am breaking this poller apart in two, with the I/O queue being polled
separately from the OS-level work in preparation for io_uring support.
When that happens, it will be necessary to know whether the poller did
work.

Signed-off-by: Glauber Costa <glauber@scylladb.com>