core: move storage aio to reactor backend
Right now the storage subsystem is placed inside the reactor.  This
works because we only ever do I/O using the linux io_submit/io_getevents
interface.

Linux now has a new, improved interface to do I/O (io_uring) that
dispatches I/O by means of a different set of system calls with a
different set of data structures.

One possible approach would be to do something similar to the
reactor_backend model, as an io_backend.

However io_uring is the most delicious piece of awesomeness I've
seen in a long while and it does much more than just I/O: the interface
can do pretty much anything, which means we should be able to replace
the other backend operations with it too.

Therefore this patch lays the ground for that to happen by moving the
storage functionality inside the reactor backend.

Signed-off-by: Glauber Costa <glauber@scylladb.com>