core: expand filesystem check at open time
We currently check whether we are opening on tmpfs and behave
accordingly.

We'd like to start taking special action for xfs too, so expand
the check. The reason as to why will become apparent in the next
patch: not all filesystems support I/O polling at the kernel level
and we want to white-list the ones that do.

Signed-off-by: Glauber Costa <glauber@scylladb.com>