io_queue: store whether or not underlying device (filesystem) supports iopoll
Newer kernels add the ability to poll for request completion without any
interrupts. This is loosely referred to as "iopoll", and the bummer is
that not all filesystems support it. Our favorite start, XFS,
obviously does. So record it if we create an I/O queue that sits on an
XFS filesystem, and pass it down to the io_request.

It must be noted that this is to some extent an oversimplification: XFS
running on an older kernel will obviously not support this. But since
we'll have to test for the presence of io_uring when setting it up, we
will later assume that any kernel that is new enough to support io_uring
is also new enough to support iopoll (io_uring actually depends on it).

Signed-off-by: Glauber Costa <glauber@scylladb.com>