io_desc: move destroy logic inside derived classs
Right now when we see an io_desc in the I/O path, we delete it.
That is because the I/O queue works this way: it creates a unique
pointer, then releases that pointer and moves the responsability
of deleting it to the completion event.

However as we try to make io_desc generic, this won't always be
the case. Maybe it comes from a fixed list, maybe it is a member
inside a structure.

Move the delete logic into the I/O queue, as the derived class is
now responsible for managing the object's lifetime.

With this change we can also make the destroyer non-virtual and
protected.

Signed-off-by: Glauber Costa <glauber@scylladb.com>